requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.1
python-dotenv==1.0.0
pydantic==2.5.0
